Permissoes

+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ use App\Models\Role;
 use App\Models\User;
 use App\Models\UserProfile;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -39,12 +40,14 @@ class UserController extends Controller
     }
 
     public function edit(User $user) {
+        Gate::authorize('edit', $user);
         $user->load(['profile', 'interests']);
         $roles = Role::all();
         return view('users.edit', compact('user', 'roles'));
     }
 
     public function update(User $user, Request $request) {
+        Gate::authorize('edit', $user);
         $input = $request->validate([
             'name' => 'required',
             'email' => 'required|email',
@@ -58,6 +61,7 @@ class UserController extends Controller
     }
 
     public function updateProfile(User $user, Request $request) {
+        Gate::authorize('edit', $user);
         $input = $request->validate([
             'type' => 'required',
             'address' => 'nullable',
@@ -69,7 +73,7 @@ class UserController extends Controller
     }
 
     public function updateInterests(User $user, Request $request) {
-
+        Gate::authorize('edit', $user);
         $input = $request->validate([
             'interests' => 'nullable|array',
         ]);
@@ -85,6 +89,7 @@ class UserController extends Controller
     }
 
     public function updateRoles(User $user, Request $request) {
+        Gate::authorize('edit', $user);
         $input = $request->validate([
             'roles' => 'required|array',
         ]);
@@ -96,6 +101,7 @@ class UserController extends Controller
     }
 
     public function destroy(User $user) {
+        Gate::authorize('destroy', $user);
         $user->delete();
         return back()->with('status', 'Usuário deletado com sucesso!');
     }
